CoMM

CoMM_S4_testing to dissecting genetic contributions to complex traits by leveraging regulatory information using both eQTL and GWAS summary statistics across whole genome

CoMM_S4_testing(stringname1, stringname2, stringname3, stringname4,
  stringname5, px, lam)

Arguments

stringname1
file for eQTL summary data file with the format described in the vignette.

stringname2
file for GWAS summary data file with the format described in the vignette.

stringname3
prefix for eQTL reference panel file with plink format (bim,bed).

stringname4
prefix for the file complementing the information of SNPs in eQTL summary statistics with the format described in the vignette

stringname5
prefix for GWAS reference panel file with plink format (bim,bed).

px
logical value(1 or 0) for indicating whether to use paramter expansion to accerate algorithm

lam
shirnkage intensity for reference panel

Value

List of model parameters

Details

CoMM_S4_testing fits the CoMM_S4 model across whole genome. It requires to provide both the eQTL and GWAS summary data, two reference panel data for eQTL and GWAS data respectively.

Author

Yi Yang, gmsyany@nus.edu.sg

Examples

file1 = "cis-eQTLs_full_chr1_1-100000.txt";
file2 = "NFBC_ph1_beta_se.txt";
file3 = "all_chr_1000G"
file4 = "cis-eQTLs_full_chr1";
file5 = "all_chr_1000G";
px = 1;
lam = 0.95;

fm = CoMM_S4_testing(file1,file2,file3, file4,file5, px, lam);

#> begin to read reference panel

#> Error in CoMM_S4_testing(file1, file2, file3, file4, file5, px, lam): Mat::col(): index out of bounds